qcacmn: Add support for 256 Client connect/disconnect event
To offload client more than 128 with IPA, update data type of
num_sap_connected to uint16_t.

fix is change the data type to uint16_t from uint8_t.

CRs-Fixed: 3717671

+++ ipa/core/inc/wlan_ipa_priv.h
@@ -706,8 +706,8 @@ struct wlan_ipa_priv {
 
 	uint8_t activated_fw_pipe;
 	uint8_t num_sap_connected;
-	uint8_t sap_num_connected_sta;
-	uint8_t sap_num_mlo_connected_sta;
+	uint16_t sap_num_connected_sta;
+	uint16_t sap_num_mlo_connected_sta;
 	uint8_t sta_connected;
 	uint32_t tx_pipe_handle;
 	uint32_t rx_pipe_handle;
